+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `make clean` command for cleaning clickgen cache
 - Build `xcursorgen` with extra flags
 - `xcursorgen.c` formated with tool **[indent](https://www.gnu.org/software/indent/)**
+- `make dev` command for development purpose
+- Linting & typing fixes inside `clickgen.builders`
 
 ### Changed
 
@@ -21,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Proper typing inheritation inside `clickgen/core.pyi`
 - Linting & Typing fixed in `XCursor` Class `clickgen/builder.py`
 - `xcursorgen/makefile` renamed to `xcursorgen/Makefile`
+- WindowsCursor support `options` instead of `args`
 
 ## [1.1.9] - 22 Mar 2021